cleanup: move `show*NotificationWithButtons()` to new `notifications.ts` (#1617)
