+++ src/commands/docker.ts
@@ -12,9 +12,9 @@ import { ResponseError } from "../clients/docker";
 import { isDockerAvailable } from "../docker/configs";
 import { LocalResourceKind } from "../docker/constants";
 import { LocalResourceWorkflow } from "../docker/workflows/base";
-import { showErrorNotificationWithButtons } from "../errors";
 import { Logger } from "../logging";
 import { ConnectionLabel } from "../models/resource";
+import { showErrorNotificationWithButtons } from "../notifications";
 import { LOCAL_DOCKER_SOCKET_PATH } from "../preferences/constants";
 import { localResourcesQuickPick } from "../quickpicks/localResources";
 import { UserEvent } from "../telemetry/events";
