+++ src/docker/workflows/cp-schema-registry.ts
@@ -8,8 +8,8 @@ import {
   HostConfig,
 } from "../../clients/docker";
 import { localSchemaRegistryConnected } from "../../emitters";
-import { showErrorNotificationWithButtons } from "../../errors";
 import { Logger } from "../../logging";
+import { showErrorNotificationWithButtons } from "../../notifications";
 import { LOCAL_KAFKA_IMAGE, LOCAL_KAFKA_IMAGE_TAG } from "../../preferences/constants";
 import { updateLocalConnection } from "../../sidecar/connections/local";
 import { UserEvent } from "../../telemetry/events";
