+++ src/quickpicks/utils/schemaSubjects.ts
@@ -1,9 +1,12 @@
 import { commands, workspace, WorkspaceConfiguration } from "vscode";
-import { DEFAULT_ERROR_NOTIFICATION_BUTTONS, showErrorNotificationWithButtons } from "../../errors";
 import { ResourceLoader } from "../../loaders";
 import { Subject } from "../../models/schema";
 import { SchemaRegistry } from "../../models/schemaRegistry";
 import { KafkaTopic } from "../../models/topic";
+import {
+  DEFAULT_ERROR_NOTIFICATION_BUTTONS,
+  showErrorNotificationWithButtons,
+} from "../../notifications";
 import { USE_TOPIC_NAME_STRATEGY } from "../../preferences/constants";
 import { SubjectNameStrategy } from "../../schemas/produceMessageSchema";
 import { schemaSubjectQuickPick, subjectNameStrategyQuickPick } from "../schemas";
