+++ src/quickpicks/utils/schemas.ts
@@ -1,9 +1,9 @@
 import { workspace, WorkspaceConfiguration } from "vscode";
-import { showErrorNotificationWithButtons } from "../../errors";
 import { ResourceLoader } from "../../loaders";
 import { Schema } from "../../models/schema";
 import { SchemaRegistry } from "../../models/schemaRegistry";
 import { KafkaTopic } from "../../models/topic";
+import { showErrorNotificationWithButtons } from "../../notifications";
 import { ALLOW_OLDER_SCHEMA_VERSIONS } from "../../preferences/constants";
 import { SubjectNameStrategy } from "../../schemas/produceMessageSchema";
 import { schemaVersionQuickPick } from "../schemas";
