+++ src/sidecar/sidecarManager.ts
@@ -23,7 +23,8 @@ import { join, normalize } from "path";
 import { Tail } from "tail";
 import { EXTENSION_VERSION, SIDECAR_OUTPUT_CHANNEL } from "../constants";
 import { observabilityContext } from "../context/observability";
-import { logError, showErrorNotificationWithButtons } from "../errors";
+import { logError } from "../errors";
+import { showErrorNotificationWithButtons } from "../notifications";
 import { SecretStorageKeys } from "../storage/constants";
 import { WriteableTmpDir } from "../utils/file";
 import { checkSidecarOsAndArch } from "./checkArchitecture";
